# `run_async` should run async