Workaround for https://github.com/surface-security/django-dkron/issues/18

This should be revisited once there are any updated there

Before change
```
➜  time ./manage.py shell -c 'from dkron.utils import run_async; run_async("shell", c="\"import time;time.sleep(10)\"")'
./manage.py shell -c   0.42s user 0.07s system 4% cpu 11.120 total

➜  time ./manage.py shell -c 'from dkron.utils import run_async; run_async("shell", c="\"import time;time.sleep(5)\"")'
./manage.py shell -c   0.43s user 0.07s system 8% cpu 6.119 total
```

Test changes in https://github.com/surface-security/django-dkron/pull/19/commits/d798b4d9eecdaa502d855af9042f767f0cf66e3d reflect the same timings, so it's not a solution.

Left with the option to use `@at +5seconds`, I'm unable to think of any other.

After changes

```
➜  time ./manage.py shell -c 'from dkron.utils import run_async; run_async("shell", c="\"import time;time.sleep(10)\"")'
./manage.py shell -c   0.43s user 0.07s system 90% cpu 0.555 total
➜  time ./manage.py shell -c 'from dkron.utils import run_async; run_async("shell", c="\"import time;time.sleep(3)\"")'
./manage.py shell -c   0.44s user 0.07s system 92% cpu 0.550 total
```

And the job is indeed triggered in dkron.