@fopina 2023-02-07
Ready for review. Note to reviewer/merger: squash commits.

I haven't squashed in the branch to keep that intermediate commit in the PR (but not worth having it in main history)