@gsilvapt 2023-02-07
One thing  we had in mind in Surface was to move `WSGI` to `ASGI` for asynchronous views. Do you think this change could impact that?