@fopina 2023-02-07
I haven’t assumed ASGI but I’d assume synchronous calls within a view will remain synchronous. And so would the HTTP itself (unless it only uses HTTP2 which I don’t think so), so it would still have the request waiting for job execution (and possibly timing out)