@gsilvapt 2023-02-07
Hi again. 

So, I was reviewing this PR and took the opportunity to set up dkron and this app locally for testing. I'm going to ask you which versions of Python and Dkron you're running because I am unable to reproduce the issue - getting a completely different issue but it's highly likely to be related to my setup: 
I ran `docker compose up` from the `dkron` repo so it's running latest everywhere, then created a venv for `django-dkron` to install deps, create super user (to access the dashboard), updated the `DKRON_URL` option to match my local `dkron` and tried the commands you quoted above. They run in less then 500ms, but I see the job gets assigned to the untriggered jobs:

<img width="1453" alt="image" src="https://user-images.githubusercontent.com/7786556/217390079-e4e0be5f-aacb-4d88-a8c1-a278dd19a0ae.png">

I'm not sure if I'm experiencing a different behavior because the jobs are not getting triggered due to some misconfiguration of the agents or the latest dkron version has something different in it. (or Python 3.11 which I doubt).

--- 
Anyway, the sole reason I wanted to test this was to try the `concurrency: allow` option, although it's not clear if that would work to start the job. or not. Inspecting the jobs this is creating without any change, seems that's the default option anyway so that shouldn't solve the problem.