@fopina 2023-02-08
Looking at the screenshot, you're missing the tag in the agents, so you're probably not running dkron with `run_dkron` as expected