@gsilvapt 2023-02-08
> Looking at the screenshot, you're missing the tag in the agents, so you're probably not running dkron with `run_dkron` as expected.
> 
> Concurrency is [documented](https://dkron.io/docs/usage/concurrency) and it is about triggering multiple excutions of same job.
> 
> > An alternative idea that came to mind while investigating dkron and the API is to maybe to keep the manual, do not runoncreate but call the corresponding API endpoint to trigger the job. Seems to be possible, although maybe not as pretty.
> 
> This was the original implementation, before `runoncreate` was added. And that's the one I mention in the PR description
> 
> > Test changes in [d798b4d](https://github.com/surface-security/django-dkron/commit/d798b4d9eecdaa502d855af9042f767f0cf66e3d) reflect the same timings, so it's not a solution.
> 
> It can be verified directly from the UI: create a job with shell `sleep 30` and click run job after. Request will be stuck for 30 secs as well. So it's not an issue with `runoncreated` but with their approach for "live running" I suppose. We'll see what comes out of the issue in dkron but in the meantime this would be a workaround

Thanks for the tip - I was pointing the app directly to the resulting dkron server out of the `docker compose` file. We should improve the app's documentation a little on how to run `run_dkron`, how to attach to existing cluster and/or configure more agents.

Anyway, I was able to reproduce the issue highlighted here. Since it's hard to get things approved upstream, I think we can proceed with these changes. We do not have a use case for it internally, AFAIK, but if we want to support async calls, then this small hack makes sense.

Thanks